Reorganize backend tests to make them more focused
